BUG: prevent DICOM receiver from interrupting segmentation process (issue #353)
- restructured and improved error handling
- moved code for preop handling into separate file
