+++ SliceTracker/SliceTrackerUtils/session.py
@@ -1,31 +1,25 @@
 import os, logging
 import vtk, ctk, ast
 import qt
-import xml
-import getpass
-import datetime
 
 import slicer
-from sessionData import SessionData, RegistrationResult, RegistrationTypeData, SegmentationData, PreopData
+from sessionData import SessionData, RegistrationResult, RegistrationTypeData
 from constants import SliceTrackerConstants
 from helpers import SeriesTypeManager
-from algorithms.automaticProstateSegmentation import AutomaticSegmentationLogic
+from preopHandler import PreopDataHandler
 
-from SlicerDevelopmentToolboxUtils.constants import DICOMTAGS, FileExtension, STYLE
+from SlicerDevelopmentToolboxUtils.constants import DICOMTAGS, STYLE
 from SlicerDevelopmentToolboxUtils.events import SlicerDevelopmentToolboxEvents
 from SlicerDevelopmentToolboxUtils.helpers import SmartDICOMReceiver
-from SlicerDevelopmentToolboxUtils.mixins import ModuleWidgetMixin, ModuleLogicMixin
-from SlicerDevelopmentToolboxUtils.exceptions import DICOMValueError, PreProcessedDataError, UnknownSeriesError
+from SlicerDevelopmentToolboxUtils.mixins import ModuleWidgetMixin
+from SlicerDevelopmentToolboxUtils.exceptions import DICOMValueError, UnknownSeriesError
 from SlicerDevelopmentToolboxUtils.widgets import IncomingDataWindow, CustomStatusProgressbar
-from SlicerDevelopmentToolboxUtils.widgets import SliceWidgetConfirmYesNoDialog
 from SlicerDevelopmentToolboxUtils.widgets import RadioButtonChoiceMessageBox
-from SlicerDevelopmentToolboxUtils.decorators import singleton
-from SlicerDevelopmentToolboxUtils.decorators import onExceptionReturnFalse, onReturnProcessEvents, onExceptionReturnNone
+from SlicerDevelopmentToolboxUtils.decorators import singleton, onExceptionReturnFalse
+from SlicerDevelopmentToolboxUtils.decorators import onReturnProcessEvents, onExceptionReturnNone
 from SlicerDevelopmentToolboxUtils.module.session import StepBasedSession
 
-
 from SliceTrackerRegistration import SliceTrackerRegistrationLogic
-from steps.plugins.segmentationValidator import SliceTrackerSegmentationValidatorPlugin
 
 
 @singleton
@@ -205,6 +199,7 @@ class SliceTrackerSession(StepBasedSession):
     self.resetAndInitializeMembers()
 
   def resetAndInitializeMembers(self):
+    self._busy = False
     self.seriesTypeManager.clear()
     self.initializeColorNodes()
     self.directory = None
@@ -243,11 +238,14 @@ class SliceTrackerSession(StepBasedSession):
   def isCurrentSeriesCoverProstateInNonPreopMode(self):
     return self.seriesTypeManager.isCoverProstate(self.currentSeries) and not self.data.usePreopData
 
+  def isBusy(self):
+    return self.isPreProcessing() or self._busy
+
   def isPreProcessing(self):
     return slicer.util.selectedModule() != self.MODULE_NAME
 
   def isCaseDirectoryValid(self):
-    return os.path.exists(self.preopDICOMDirectory) and os.path.exists(self.intraopDICOMDirectory)
+    return all(os.path.exists(p) for p in [self.preopDICOMDirectory, self.intraopDICOMDirectory])
 
   def isRunning(self):
     return not self.directory in [None, '']
@@ -256,12 +254,12 @@ class SliceTrackerSession(StepBasedSession):
     self.newCaseCreated = getattr(self, "newCaseCreated", False)
     if self.newCaseCreated:
       return
-    if not self.directory or not self.isCaseDirectoryValid():
-      slicer.util.warningDisplay("The selected case directory seems not to be valid", windowTitle="SliceTracker")
-      self.close(save=False)
-    else:
+    if self.isCaseDirectoryValid():
       self.loadCaseData()
       self.invokeEvent(self.CaseOpenedEvent)
+    else:
+      slicer.util.warningDisplay("The selected case directory seems not to be valid", windowTitle="SliceTracker")
+      self.close(save=False)
 
   def createNewCase(self, destination):
     self.newCaseCreated = True
@@ -301,10 +299,9 @@ class SliceTrackerSession(StepBasedSession):
                                         % ("completed" if completed else "started",
                                            "open it" if completed else "continue this session")):
       slicer.app.layoutManager().blockSignals(True)
-      self._loading = True
+      self._busy = True
       self.data.load(filename)
       self.postProcessLoadedSessionData()
-      self._loading = False
       slicer.app.layoutManager().blockSignals(False)
       self.invokeEvent(self.LoadingMetadataSuccessfulEvent)
     else:
@@ -363,6 +360,7 @@ class SliceTrackerSession(StepBasedSession):
     preopDataManager.handle()
 
   def onPreprocessingStarted(self, caller, event):
+    self._busy = True
     self.startIntraopDICOMReceiver()
 
   def resetPreopDICOMReceiver(self):
@@ -374,7 +372,7 @@ class SliceTrackerSession(StepBasedSession):
 
   def startIntraopDICOMReceiver(self):
     self.resetPreopDICOMReceiver()
-    logging.info("Starting DICOM Receiver for intra-procedural data")
+    logging.debug("Starting DICOM Receiver for intra-procedural data")
     if not self.data.completed:
       self.resetIntraopDICOMReceiver()
       self.intraopDICOMReceiver = SmartDICOMReceiver(self.intraopDICOMDirectory,
@@ -457,9 +455,9 @@ class SliceTrackerSession(StepBasedSession):
 
   def getPatientIDValidationSource(self):
     if len(self.loadableList.keys()) > 1:
-      keylist = self.loadableList.keys()
-      keylist.sort(key=lambda x: RegistrationResult.getSeriesNumberFromString(x))
-      return self.loadableList[keylist[0]][0]
+      keys = self.loadableList.keys()
+      keys.sort(key=lambda x: RegistrationResult.getSeriesNumberFromString(x))
+      return self.loadableList[keys[0]][0]
     else:
       return None
 
@@ -467,7 +465,6 @@ class SliceTrackerSession(StepBasedSession):
     try:
       volume = self.alreadyLoadedSeries[series]
     except KeyError:
-      logging.info("Need to load volume")
       files = self.loadableList[series]
       loadables = self.scalarVolumePlugin.examine([files])
       assert len(loadables)
@@ -526,28 +523,34 @@ class SliceTrackerSession(StepBasedSession):
     return None
 
   def loadCaseData(self):
-    if not os.path.exists(os.path.join(self.outputDirectory, SliceTrackerConstants.JSON_FILENAME)):
+    self._busy = True
+    if self.hasJSONResults():
+      self.load()
+    else:
       if PreopDataHandler.wasDirectoryPreprocessed(self.preprocessedDirectory):
+        self.startIntraopDICOMReceiver()
         preopDataManager = self.createPreopHandler()
         preopDataManager.loadPreProcessedData()
       else:
-        if len(os.listdir(self.preopDICOMDirectory)):
-          self.onPreopDataReceptionFinished()
-        elif len(os.listdir(self.intraopDICOMDirectory)):
-          self.data.usePreopData = False
-          self.startIntraopDICOMReceiver()
-        else:
-          self.startPreopDICOMReceiver()
+        self.continueWithUnprocessedData()
+
+  def continueWithUnprocessedData(self):
+    if os.listdir(self.preopDICOMDirectory):
+      self.onPreopDataReceptionFinished()
+    elif os.listdir(self.intraopDICOMDirectory):
+      self.data.usePreopData = False
+      self.startIntraopDICOMReceiver()
     else:
-      self.openSavedSession()
+      self.startPreopDICOMReceiver()
+
+  def hasJSONResults(self):
+    return os.path.exists(os.path.join(self.outputDirectory, SliceTrackerConstants.JSON_FILENAME))
 
   def onPreprocessingSuccessful(self, caller, event):
     self.setupPreopLoadedTargets()
+    self._busy = False
     self.invokeEvent(self.PreprocessingSuccessfulEvent)
     self.startIntraopDICOMReceiver()
-
-  def openSavedSession(self):
-    self.load()
 
   def setupPreopLoadedTargets(self):
     targets = self.data.initialTargets
@@ -623,10 +626,6 @@ class SliceTrackerSession(StepBasedSession):
     seriesType = self.seriesTypeManager.getSeriesType(series)
     return not self.isAnyListItemInString(seriesType,[self.getSetting("COVER_PROSTATE_PATTERN"),
                                                       self.getSetting("COVER_TEMPLATE_PATTERN")])
-
-  def isLoading(self):
-    self._loading = getattr(self, "_loading", False)
-    return self._loading
 
   def takeActionForCurrentSeries(self):
     event = None
@@ -744,8 +743,7 @@ class SliceTrackerSession(StepBasedSession):
 
   def onRegistrationResultStatusChanged(self, caller, event):
     self.skipAllUnregisteredPreviousSeries(self.currentResult.name)
-    self._loading = getattr(self, "_loading", None)
-    if self._loading:
+    if self._busy:
       return
     self.save()
     self.invokeEvent(self.RegistrationStatusChangedEvent)
@@ -780,337 +778,3 @@ class SliceTrackerSession(StepBasedSession):
 
   def _getConsent(self):
     return RadioButtonChoiceMessageBox("Who gave consent?", options=["Clinician", "Operator"]).exec_()
-
-
-class PreprocessedDataHandlerBase(ModuleWidgetMixin, ModuleLogicMixin):
-
-  PreprocessingStartedEvent = vtk.vtkCommand.UserEvent + 434
-  PreprocessingFinishedEvent = vtk.vtkCommand.UserEvent + 435
-  PreprocessedDataErrorEvent = vtk.vtkCommand.UserEvent + 436
-
-  @staticmethod
-  def wasDirectoryPreprocessed(directory):
-    raise NotImplementedError
-
-  @property
-  def outputDirectory(self):
-    self._outputDirectory = getattr(self, "_outputDirectory", None)
-    return self._outputDirectory
-
-  @outputDirectory.setter
-  def outputDirectory(self, directory):
-    if directory and not os.path.exists(directory):
-      self.createDirectory(directory)
-    self._outputDirectory = directory
-
-  def __init__(self, inputDirectory, outputDirectory):
-    self._inputDirectory = inputDirectory
-    self.outputDirectory = outputDirectory
-
-  def cleanup(self):
-    pass
-
-  def handle(self):
-    raise NotImplementedError
-
-
-class PreopDataHandler(PreprocessedDataHandlerBase):
-
-  MODULE_NAME = SliceTrackerConstants.MODULE_NAME
-
-  @property
-  def segmentationData(self):
-    try:
-      return self.data.preopData.segmentation
-    except AttributeError:
-      return None
-
-  @segmentationData.setter
-  def segmentationData(self, value):
-    assert self.preopData
-    self.preopData.segmentation = value
-
-  @property
-  def preopData(self):
-    return self.data.preopData
-
-  @preopData.setter
-  def preopData(self, value):
-    self.data.preopData = value
-
-  @staticmethod
-  def getFirstMpReviewPreprocessedStudy(directory):
-    # TODO add check here and selected the one which has targets in it
-    # TODO: if several studies are available provide a drop down or anything similar for choosing
-    directoryNames = filter(os.path.isdir, [os.path.join(directory, f) for f in os.listdir(directory)])
-    return None if not len(directoryNames) else directoryNames[0]
-
-  @staticmethod
-  def wasDirectoryPreprocessed(directory):
-    # TODO: this check needs to be more specific with expected sub directories
-    firstDir = PreopDataHandler.getFirstMpReviewPreprocessedStudy(directory)
-    if not firstDir:
-      return False
-    from mpReview import mpReviewLogic
-    return mpReviewLogic.wasmpReviewPreprocessed(directory)
-
-  def __init__(self, inputDirectory, outputDirectory, data):
-    super(PreopDataHandler, self).__init__(inputDirectory, outputDirectory)
-    self.data = data
-
-  def __del__(self):
-    super(PreopDataHandler, self).__del__()
-
-  def handle(self):
-    self.invokeEvent(self.PreprocessingStartedEvent)
-    if self._runPreProcessing():
-      self.runModule(invokeEvent=False)
-    else:
-      slicer.util.infoDisplay("No DICOM data could be processed. Please select another directory.",
-                              windowTitle="SliceTracker")
-
-  def _runPreProcessing(self):
-    from mpReviewPreprocessor import mpReviewPreprocessorLogic
-    mpReviewPreprocessorLogic = mpReviewPreprocessorLogic()
-    progress = self.createProgressDialog()
-    progress.canceled.connect(lambda : mpReviewPreprocessorLogic.cancelProcess())
-
-    @onReturnProcessEvents
-    def updateProgressBar(**kwargs):
-      for key, value in kwargs.iteritems():
-        if hasattr(progress, key):
-          setattr(progress, key, value)
-
-    success = mpReviewPreprocessorLogic.importAndProcessData(self._inputDirectory, outputDir=self.outputDirectory,
-                                                             copyDICOM=False,
-                                                             progressCallback=updateProgressBar)
-    progress.canceled.disconnect(lambda : mpReviewPreprocessorLogic.cancelProcess())
-    progress.close()
-    return success
-
-  def runModule(self, invokeEvent=True):
-
-    def onModuleReturn():
-      slicer.modules.mpReviewWidget.saveButton.clicked.disconnect(onModuleReturn)
-      self.layoutManager.selectModule(self.MODULE_NAME)
-      slicer.mrmlScene.Clear(0)
-      self.loadPreProcessedData()
-
-    self.setSetting('InputLocation', None, moduleName="mpReview")
-    self.layoutManager.selectModule("mpReview")
-    mpReview = slicer.modules.mpReviewWidget
-    self.setSetting('InputLocation', self.outputDirectory, moduleName="mpReview")
-    mpReview.onReload()
-    slicer.modules.mpReviewWidget.saveButton.clicked.connect(onModuleReturn)
-    self.layoutManager.selectModule(mpReview.moduleName)
-    if invokeEvent:
-      self.invokeEvent(self.PreprocessingStartedEvent)
-
-  def loadPreProcessedData(self):
-    try:
-      self.loadMpReviewProcessedData()
-    except PreProcessedDataError:
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-
-  def loadMpReviewProcessedData(self):
-    studyDir = self.getFirstMpReviewPreprocessedStudy(self.outputDirectory)
-    resourcesDir = os.path.join(studyDir, 'RESOURCES')
-    if not self.isMpReviewStudyDirectoryValid(resourcesDir):
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-      return
-
-    self.findPreopImageAndSegmentationPaths(resourcesDir)
-
-    message = None
-    if not self.preopSegmentationPath:
-      slicer.util.errorDisplay("No eligible series found for preop AX T2 segmentation. MpReview might not have "
-                               "processed the right series or series names are different from BWH internally used ones")
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-      return
-
-    self.data.initialTargetsPath = os.path.join(studyDir, 'Targets')
-
-    loadedPreopVolume = self.loadPreopVolume()
-    loadedPreopTargets = self.loadPreopTargets()
-    loadedPreopT2Label = self.loadT2Label() if os.path.exists(self.preopSegmentationPath) else False
-
-    if message or not (loadedPreopT2Label and loadedPreopVolume and loadedPreopTargets):
-      if loadedPreopTargets and loadedPreopVolume and \
-          str(self.getSetting("Use_Deep_Learning", moduleName=self.MODULE_NAME)).lower() == "true":
-        if slicer.util.confirmYesNoDisplay("No WholeGland segmentation found in preop data. Automatic segmentation is "
-                                           "available. Would you like to proceed with the automatic segmentation?",
-                                           windowTitle="SliceTracker"):
-          self._createPreopData(algorithm="Automatic")
-          self.runAutomaticSegmentation()
-        else:
-          self.onPreopLoadingFailed()
-      else:
-        self.onPreopLoadingFailed()
-    else:
-      if not self.data.usePreopData:
-        self._createPreopData(algorithm="Manual")
-        self.segmentationData.note = "mpReview preprocessed"
-        self.segmentationData._label = self.data.initialLabel
-      self.invokeEvent(self.PreprocessingFinishedEvent)
-
-  def _createPreopData(self, algorithm, segmentationType="Prostate"):
-    self.preopData = PreopData()
-    self.segmentationData = SegmentationData(segmentationType=segmentationType, algorithm=algorithm)
-
-  def runAutomaticSegmentation(self):
-    logic = AutomaticSegmentationLogic()
-    logic.addEventObserver(logic.DeepLearningFinishedEvent, self.onSegmentationFinished)
-    logic.addEventObserver(logic.DeepLearningFailedEvent, self.onPreopLoadingFailed)
-    customStatusProgressBar = CustomStatusProgressbar()
-    customStatusProgressBar.text = "Running DeepInfer for automatic prostate segmentation"
-
-    from mpReview import mpReviewLogic
-    mpReviewColorNode, _ = mpReviewLogic.loadColorTable(self.getSetting("Color_File_Name", moduleName=self.MODULE_NAME))
-
-    domain = 'BWH_WITHOUT_ERC'
-    prompt = SliceWidgetConfirmYesNoDialog(self.data.initialVolume,
-                                           "Was an endorectal coil used during preop acquisition?").exec_()
-
-    self.preopData.usedERC = False
-    if prompt == qt.QDialogButtonBox.Yes:
-      self.preopData.usedERC = True
-      domain = 'BWH_WITH_ERC'
-    elif prompt == qt.QDialogButtonBox.Cancel:
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-      return
-
-    self.segmentationData.startTime = self.getTime()
-    logic.run(self.data.initialVolume, domain, mpReviewColorNode)
-
-  @vtk.calldata_type(vtk.VTK_OBJECT)
-  def onSegmentationValidated(self, caller, event, labelNode):
-    if "_modified" in labelNode.GetName():
-      self.segmentationData.userModified["endTime"] = self.getTime()
-      self.segmentationData._modifiedLabel = labelNode
-    else:
-      self.segmentationData.userModified = None
-    if not self.preopSegmentationPath:
-      self.createDirectory(self.preopSegmentationPath)
-    segmentedColorName = self.getSetting("Segmentation_Color_Name", moduleName=self.MODULE_NAME)
-    timestamp = datetime.datetime.now().strftime("%Y%m%d%H%M%S")
-    filename = getpass.getuser() + '-' + segmentedColorName + '-' + timestamp
-    self.saveNodeData(labelNode, self.preopSegmentationPath, extension=FileExtension.NRRD, name=filename)
-    self.loadPreProcessedData()
-    customStatusProgressBar = CustomStatusProgressbar()
-    customStatusProgressBar.text = "Done automatic prostate segmentation"
-
-  @vtk.calldata_type(vtk.VTK_OBJECT)
-  def onSegmentationFinished(self, caller, event, labelNode):
-    # TODO: this mode should block new received images to be processed
-    self.segmentationData.endTime = self.getTime()
-    self.segmentationData._label = labelNode
-    segmentationValidator = SliceTrackerSegmentationValidatorPlugin(self.data.initialVolume, labelNode)
-    segmentationValidator.addEventObserver(segmentationValidator.ModifiedEvent, self.onSegmentationModificationStarted)
-    segmentationValidator.addEventObserver(segmentationValidator.FinishedEvent, self.onSegmentationValidated)
-    segmentationValidator.addEventObserver(segmentationValidator.CanceledEvent, self.onSegmentationValidationFailed)
-    segmentationValidator.run()
-
-  def onSegmentationModificationStarted(self, caller, event):
-    self.segmentationData.setModified(startTime=self.getTime())
-
-  def onSegmentationValidationFailed(self, caller=None, event=None):
-    message = "Segmentation validation failed.\nWholeGland segmentation is require to proceed." \
-              "\n\nDo you want to " \
-              "open/revisit pre-processing for the current case?"
-    if slicer.util.confirmYesNoDisplay(message, windowTitle="SliceTracker"):
-      # TODO: start intraop without listening to signals
-      self.runModule()
-    else:
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-
-  def isMpReviewStudyDirectoryValid(self, resourcesDir):
-    logging.debug(resourcesDir)
-    if not os.path.exists(resourcesDir):
-      logging.info("The selected directory does not fit the mpReview directory structure. Make sure that you select "
-                   "the study root directory which includes directories RESOURCES")
-      return False
-    return True
-
-  def onPreopLoadingFailed(self, caller=None, event=None):
-    message = "Loading preop data failed.\nMake sure that the correct mpReview directory structure is used." \
-              "\n\nSliceTracker expects a T2 volume, WholeGland segmentation and target(s). Do you want to " \
-              "open/revisit pre-processing for the current case?"
-    if slicer.util.confirmYesNoDisplay(message, windowTitle="SliceTracker"):
-      # TODO: start intraop without listening to signals
-      self.runModule()
-    else:
-      self.invokeEvent(self.PreprocessedDataErrorEvent)
-
-  def findPreopImageAndSegmentationPaths(self, resourcesDir):
-    from mpReview import mpReviewLogic
-    seriesMap, metaFile = mpReviewLogic.loadMpReviewProcessedData(resourcesDir)
-
-    self.preopImagePath = None
-    self.preopSegmentationPath = None
-    # segmentedColorName = self.getSetting("Segmentation_Color_Name", moduleName=self.MODULE_NAME) # TODO: correct?
-
-    for series in seriesMap:
-      seriesName = str(seriesMap[series]['LongName'])
-      logging.debug('series Number ' + series + ' ' + seriesName)
-
-      imagePath = os.path.join(seriesMap[series]['NRRDLocation'])
-
-      xmlFile = imagePath.replace(".nrrd", ".xml")
-      dom = xml.dom.minidom.parse(xmlFile)
-      seriesDescription = self.findElement(dom, "SeriesDescription")
-
-      segmentationsPath = os.path.join(os.path.dirname(os.path.dirname(imagePath)), 'Segmentations')
-
-      import re
-      regex = self.getSetting("PLANNING_IMAGE_PATTERN", moduleName=self.MODULE_NAME)
-
-      if re.match(regex, seriesDescription) or seriesDescription == regex:
-        logging.debug(' FOUND THE SERIES OF INTEREST, ITS ' + seriesName)
-        logging.debug(' LOCATION OF VOLUME : ' + str(seriesMap[series]['NRRDLocation']))
-        logging.debug(' LOCATION OF IMAGE path : ' + str(imagePath))
-        logging.debug(' EXPECTED LOCATION OF SEGMENTATION path : ' + segmentationsPath)
-
-        self.preopImagePath = seriesMap[series]['NRRDLocation']
-        self.preopSegmentationPath = segmentationsPath
-        break
-
-  def loadT2Label(self):
-    if self.data.initialLabel and self.data.initialLabel.GetScene() is not None:
-      return True
-    mostRecentFilename = self.getMostRecentWholeGlandSegmentation(self.preopSegmentationPath)
-    success = False
-    if mostRecentFilename:
-      filename = os.path.join(self.preopSegmentationPath, mostRecentFilename)
-      success, self.data.initialLabel = slicer.util.loadLabelVolume(filename, returnNode=True)
-      if success:
-        self.data.initialLabel.SetName('t2-label')
-    return success
-
-  def loadPreopVolume(self):
-    if self.data.initialVolume and self.data.initialVolume.GetScene() is not None:
-      return True
-    success, self.data.initialVolume = slicer.util.loadVolume(self.preopImagePath, returnNode=True)
-    if success:
-      self.data.initialVolume.SetName('VOLUME-PREOP')
-    return success
-
-  def loadPreopTargets(self):
-    if self.data.initialTargets and self.data.initialTargets.GetScene() is not None:
-      return True
-    if not os.path.exists(self.data.initialTargetsPath):
-      return False
-    mostRecentTargets = self.getMostRecentTargetsFile(self.data.initialTargetsPath)
-    success = False
-    if mostRecentTargets:
-      filename = os.path.join(self.data.initialTargetsPath, mostRecentTargets)
-      success, self.data.initialTargets = slicer.util.loadMarkupsFiducialList(filename, returnNode=True)
-      if success:
-        self.data.initialTargets.SetName('initialTargets')
-        self.data.initialTargets.SetLocked(True)
-    return success
-
-  def getMostRecentWholeGlandSegmentation(self, path):
-    return self.getMostRecentFile(path, FileExtension.NRRD, filter="WholeGland")
-
-  def getMostRecentTargetsFile(self, path):
-    return self.getMostRecentFile(path, FileExtension.FCSV)
